requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2